requirements.txt
PyYAML
tkextrafont
fonttools
ttkthemes